requirements.txt
setuptools~=50.3.2
mysql-connector-python~=8.0.22